backend/requirements.txt
flask
flask-cors
opencv-python
ultralytics
paddleocr
paddlepaddle
pymongo
numpy
python-dotenv
gunicorn
